Move to util package
